biosignal-service/requirements.txt
# Amazon Chronos and ML dependencies
# Using CPU-only torch for smaller image size
--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.0.0
transformers>=4.35.0
chronos-forecasting>=1.0.0

# Scientific computing
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.0.0

# Flask API
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.0.0

# Utilities
python-dotenv>=1.0.0
redis>=5.0.0

# Development
pytest>=7.4.0
matplotlib>=3.8.0
